requirements.txt
# === Core LLM and LangChain ===
openai>=1.43.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.1.0


# === Vector Store ===
chromadb>=0.5.3

# === PDF and data handling ===
pypdf>=4.2.0
tiktoken>=0.7.0
numpy>=1.26.0
pandas>=2.2.0
langchain-text-splitters>=0.3.0

# === UI ===
gradio>=4.44.0

# === Config & Utilities ===
python-dotenv>=1.0.1
rich>=13.7.0

# === Optional (for multi-agent extensions / async orchestration) ===
tenacity>=8.3.0
requests>=2.32.0

